+++ src/main/java/chess/domain/chesspiece/Bishop.java
@@ -1,21 +1,21 @@
-package chess.domain.chessPiece;
+package chess.domain.chesspiece;
 
 import chess.domain.position.Position;
 
-import static chess.domain.chessPiece.Role.*;
+import static chess.domain.chesspiece.Role.BLACK_BISHOP;
+import static chess.domain.chesspiece.Role.WHITE_BISHOP;
 
-public class King extends Piece {
+public class Bishop extends Piece {
 
-    public King(Team team) {
+    public Bishop(Team team) {
         super(team);
     }
 
-    //TODO : 리팩토링
     @Override
     protected void validateMovingRule(Position source, Position target) {
         int rowDistance = source.calculateRowDistance(target);
         int columnDistance = source.calculateColumnDistance(target);
-        if (rowDistance != 1 && columnDistance != 1) {
+        if (rowDistance != columnDistance) {
             throw new IllegalArgumentException("이동할 수 없습니다.");
         }
     }
@@ -23,8 +23,8 @@ public class King extends Piece {
     @Override
     public Role getRole() {
         if (team.isWhite()) {
-            return WHITE_KING;
+            return WHITE_BISHOP;
         }
-        return BLACK_KING;
+        return BLACK_BISHOP;
     }
 }
