+++ src/main/java/chess/domain/chesspiece/King.java
@@ -1,20 +1,21 @@
-package chess.domain.chessPiece;
+package chess.domain.chesspiece;
 
 import chess.domain.position.Position;
 
-import static chess.domain.chessPiece.Role.*;
+import static chess.domain.chesspiece.Role.*;
 
-public class Queen extends Piece {
+public class King extends Piece {
 
-    public Queen(Team team) {
+    public King(Team team) {
         super(team);
     }
 
+    //TODO : 리팩토링
     @Override
     protected void validateMovingRule(Position source, Position target) {
         int rowDistance = source.calculateRowDistance(target);
         int columnDistance = source.calculateColumnDistance(target);
-        if (source.isDifferentRow(target) && source.isDifferentColumn(target) && rowDistance != columnDistance) {
+        if (rowDistance != 1 && columnDistance != 1) {
             throw new IllegalArgumentException("이동할 수 없습니다.");
         }
     }
@@ -22,8 +23,8 @@ public class Queen extends Piece {
     @Override
     public Role getRole() {
         if (team.isWhite()) {
-            return WHITE_QUEEN;
+            return WHITE_KING;
         }
-        return BLACK_QUEEN;
+        return BLACK_KING;
     }
 }
