+++ src/main/java/chess/domain/chesspiece/Role.java
@@ -1,4 +1,4 @@
-package chess.domain.chessPiece;
+package chess.domain.chesspiece;
 
 public enum Role {
     BLACK_KING,
