+++ src/main/java/chess/domain/position/Position.java
@@ -1,6 +1,6 @@
 package chess.domain.position;
 
-import chess.domain.chessPiece.Team;
+import chess.domain.chesspiece.Team;
 
 import java.util.Objects;
